third_party/blink/web_tests/fast/forms/time-multiple-fields/time-multiple-fields-open-picker-key-bindings-expected.txt

CONSOLE WARNING: The specified value "2011-05-01" does not conform to the required format.  The format is "HH:mm", "HH:mm:ss" or "HH:mm:ss.SSS" where HH is 00-23, mm is 00-59, ss is 00-59, and SSS is 000-999.
F4 or Alt+Down should not open the calendar picker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("mock-page-popup") is null
PASS document.getElementById("mock-page-popup") is null
PASS successfullyParsed is true

TEST COMPLETE
